Remove /Resources (See desc)
Resources contains all of the API keys, message logs, etc. for Kaguya. Considering that I am not supporting self hosting, I have decided to remove this folder from the Github. If you wish to make a fork, when you run the .exe, the bot will automatically create a /Resources folder for you with all of the necessary files, so don't worry.