Got docker working completely. Targeting .NET Core 3.1
Kaguya is now able to be built via docker commands and is targeted to run on the .NET Core 3.1 framework.

Next update will be making changes to some file structures (such as where the Resources folder is built) so that we can use them with .NET CLI commands + Docker builds. This would allow Kaguya to be fully functioning via these utilities.