Fixed bug with logging command history.
When a command is successfully executed, the object inserted into the database will now have the actual name of the command instead of the content of the message. This is crucial for accurately logging which command was used and to see how many times a particular command has been executed! :D